Extraced RpcHandler implementation
